Update all snippets
Update all the current snippets
