+++ docs/csharp/event-pattern.md
@@ -32,7 +32,7 @@ Using an event model provides some design advantages. You can create multiple ev
 
 Here's the initial event argument declaration for finding a sought file:
 
-[!code-csharp[EventArgs](../../samples/snippets/csharp/events/Program.cs#EventArgsV1 "Define event arguments")]
+:::code language="csharp" source="./snippets/events/VersionOne.cs" id="EventArgsV1":::
 
 Even though this type looks like a small, data-only type, you should follow the convention and make it a reference (`class`) type. That means the argument object is passed by reference, and any updates to the data are viewed by all subscribers. The first version is an immutable object. You should prefer to make the properties in your event argument type immutable. That way, one subscriber can't change the values before another subscriber sees them. (There are exceptions to this practice, as you see later.)
 
@@ -40,21 +40,21 @@ Next, we need to create the event declaration in the FileSearcher class. Using t
 
 Let's fill out the FileSearcher class to search for files that match a pattern, and raise the correct event when a match is discovered.
 
-[!code-csharp[FileSearcher](../../samples/snippets/csharp/events/Program.cs#FileSearcherV1 "Create the initial file searcher")]
+:::code language="csharp" source="./snippets/events/VersionOne.cs" id="FileSearcherV1":::
 
 ## Define and raise field-like events
 
 The simplest way to add an event to your class is to declare that event as a public field, as in the preceding example:
 
-[!code-csharp[DeclareEvent](../../samples/snippets/csharp/events/Program.cs#DeclareEvent "Declare the file found event")]
+:::code language="csharp" source="./snippets/events/Program.cs" id="DeclareEvent":::
 
 This looks like it's declaring a public field, which would appear to be a bad object-oriented practice. You want to protect data access through properties, or methods. While this code might look like a bad practice, the code generated by the compiler does create wrappers so that the event objects can only be accessed in safe ways. The only operations available on a field-like event are *add handler*:
 
-[!code-csharp[DeclareEventHandler](../../samples/snippets/csharp/events/Program.cs#DeclareEventHandler "Declare the file found event handler")]
+:::code language="csharp" source="./snippets/events/Program.cs" id="AttachEventHandler":::
 
 And *remove handler*:
 
-[!code-csharp[RemoveEventHandler](../../samples/snippets/csharp/events/Program.cs#RemoveHandler "Remove the event handler")]
+:::code language="csharp" source="./snippets/events/Program.cs" id="DetachHandler":::
 
 There's a local variable for the handler. If you used the body of the lambda, the remove wouldn't work correctly. It would be a different instance of the delegate, and silently do nothing.
 
@@ -77,42 +77,17 @@ The second pattern would only cancel the operation if all subscribers wanted the
 
 Let's implement the first version for this sample. You need to add a boolean field named `CancelRequested` to the `FileFoundArgs` type:
 
-[!code-csharp[EventArgs](../../samples/snippets/csharp/events/Program.cs#EventArgs "Update event arguments")]
+:::code language="csharp" source="./snippets/events/Program.cs" id="EventArgs":::
 
 This new field is automatically initialized to `false` so you don't cancel accidentally. The only other change to the component is to check the flag after raising the event to see if any of the subscribers requested a cancellation:
 
-```csharp
-private void SearchDirectory(string directory, string searchPattern)
-{
-    foreach (var file in Directory.EnumerateFiles(directory, searchPattern))
-    {
-        FileFoundArgs args = RaiseFileFound(file);
-        if (args.CancelRequested)
-        {
-            break;
-        }
-    }
-}
-
-private FileFoundArgs RaiseFileFound(string file)
-{
-    var args = new FileFoundArgs(file);
-    FileFound?.Invoke(this, args);
-    return args;
-}
-```
+:::code language="csharp" source="./snippets/events/Program.cs" id="SearchWithCancel":::
 
 One advantage of this pattern is that it isn't a breaking change. None of the subscribers requested cancellation before, and they still aren't. None of the subscriber code requires updates unless they want to support the new cancel protocol.
 
 Let's update the subscriber so that it requests a cancellation once it finds the first executable:
 
-```csharp
-EventHandler<FileFoundArgs> onFileFound = (sender, eventArgs) =>
-{
-    Console.WriteLine(eventArgs.FoundFile);
-    eventArgs.CancelRequested = true;
-};
-```
+:::code language="csharp" source="./snippets/events/Program.cs" id="CancelSearch":::
 
 ## Adding another event declaration
 
@@ -122,25 +97,25 @@ This method could get to be a lengthy operation in a directory with many subdire
 
 You start by creating the new EventArgs derived class for reporting the new directory and progress.
 
-[!code-csharp[DirEventArgs](../../samples/snippets/csharp/events/Program.cs#SearchDirEventArgs "Define search directory event arguments")]
+:::code language="csharp" source="./snippets/events/Program.cs" id="SearchDirEventArgs":::
 
 Again, you can follow the recommendations to make an immutable reference type for the event arguments.
 
 Next, define the event. This time, you use a different syntax. In addition to using the field syntax, you can explicitly create the event property with add and remove handlers. In this sample, you don't need extra code in those handlers, but this shows how you would create them.
 
-[!code-csharp[Declare event with add and remove handlers](../../samples/snippets/csharp/events/Program.cs#DeclareSearchEvent "Declare the event with add and remove handlers")]
+:::code language="csharp" source="./snippets/events/Program.cs" id="DeclareSearchEvent":::
 
 In many ways, the code you write here mirrors the code the compiler generates for the field event definitions you saw earlier. You create the event using syntax similar to  [properties](programming-guide/classes-and-structs/properties.md). Notice that the handlers have different names: `add` and `remove`. These accessors are called to subscribe to the event, or unsubscribe from the event. Notice that you also must declare a private backing field to store the event variable. This variable is initialized to null.
 
 Next, let's add the overload of the `Search` method that traverses subdirectories and raises both events. The easiest way is to use a default argument to specify that you want to search all directories:
 
-[!code-csharp[SearchImplementation](../../samples/snippets/csharp/events/Program.cs#FinalImplementation "Implementation to search directories")]
+:::code language="csharp" source="./snippets/events/Program.cs" id="FinalImplementation":::
 
 At this point, you can run the application calling the overload for searching all subdirectories. There are no subscribers on the new `DirectoryChanged` event, but using the `?.Invoke()` idiom ensures it works correctly.
 
 Let's add a handler to write a line that shows the progress in the console window.
 
-[!code-csharp[Search](../../samples/snippets/csharp/events/Program.cs#Search "Declare event handler")]
+:::code language="csharp" source="./snippets/events/Program.cs" id="Search":::
 
 You saw patterns that are followed throughout the .NET ecosystem. By learning these patterns and conventions, you're writing idiomatic C# and .NET quickly.
